Build of knex query class to interface with database api
Continuation of the knex query class at api/query_class
the purpose of this class is to simplify queries from the client to server
as they are all ajax calls.
 This allows us to format a request to the server
via a resful route with paramaters either in the req body in ajax posts or
in the req query in ajax gets.

With a formatted response the query class will direct the request to the appropriate
knex crud function and return the results via callback.
further testing required