Changed migrations and related logic on api routes
Modified migrations as follows:
 **we were having issues with our restful routes to the database especially with tables that had
foreign key constraints**
 TO solve this, it became apparent exploding the database and starting from scratch would be easier
than tiddling with migrations.

in Users Table:  We added a unique constrain to user email, and a created_at timestamp
in Goals Table:  We added a onDeleteCascade constraint to the userId foreing key and create_at timestamp
	and removed challenge_id from the table(challenges table is now responsible keeping track of two goals)
in Tasks Table:  We added a onDeleteCascade constraint to the goalsId foreign key and created_at timestamp

In seeds: modified goals seeds to reflect changes (removed challenge_id)

Because of the Users email unique constraint; our /api/users/create route was having breaking
if a user tried to test this constraint. We wrote a simple conditional for now to avoid breaking the server
.

All of these chnges have been tested and are working so far